- Fixed all merge conflicts
- Renamed spl_query to sql_query (typo fix per review)
- Removed allow_experimental_correlated_subqueries setting (already default)
- Kept both postgresql_dotnet_client and mysql_dotnet_client in cluster.py
- Updated executeQuery call to match new signature

+++ src/BridgeHelper/LibraryBridgeHelper.cpp
@@ -1,9 +1,10 @@
-#include "LibraryBridgeHelper.h"
+#include <BridgeHelper/LibraryBridgeHelper.h>
 
 #include <Core/ServerSettings.h>
 #include <Core/Settings.h>
 #include <Common/ShellCommandsHolder.h>
 #include <IO/ConnectionTimeouts.h>
+#include <Interpreters/Context.h>
 
 namespace DB
 {
@@ -38,7 +39,7 @@ Poco::URI LibraryBridgeHelper::createBaseURI() const
 {
     Poco::URI uri;
     uri.setHost(bridge_host);
-    uri.setPort(bridge_port);
+    uri.setPort(static_cast<uint16_t>(bridge_port));
     uri.setScheme("http");
     return uri;
 }
